refactor(memory): port HTTPEmbedder into memory + wire semantic search (#337)
Step 1.3 sub-part B: move HTTPEmbedder out of memstore and hand the same
embedder instance to both the legacy memstore and the unified memory.Store
so memory.Search uses the vec0 path in production.

- Move internal/memstore/http_embedder{,_test}.go → internal/memory/.
  Package rename only; the implementation is byte-for-byte the same.
  The interface-compliance test now asserts against memory.Embedder.
- Keep memstore.EmbedderI, memstore.HTTPEmbedder, memstore.NewHTTPEmbedder
  as thin aliases over the memory types so in-flight memstore callers
  compile unchanged during the transition. These aliases will be removed
  once memstore itself is retired.
- In cmd/alf-daemon/main.go: resolve the embedder once at the shared
  scope and pass it to both memstore.New and memory.NewSQLiteStore via
  memory.WithEmbedder. Gating stays on cfg.EffectiveMemoryEnabled() so
  the embedder is not spawned when memory is disabled.
- resolveEmbedder / startHTTPEmbedder now reference memory.Embedder and
  *memory.HTTPEmbedder directly.

Result: memory.Store.Index/Search are wired end-to-end with the HTTP
embed-server in production — no caller of memstore.Store has moved yet
(sub-ticket C), but the memory path is live for anyone who opens the
unified store.
